refact: rename types folder to domain
